Based on previous feedback and suggestions, the following changes have been made:
1. Fixed incorrect code issues that were previously introduced
2. Removed redundant imports and variable declarations
3. Consolidated three isProblematic methods into DeviceUtils and updated all references

Changes:
- Added isProblematicAdrenoGpu() method to DeviceUtils.kt to provide a unified
  implementation for detecting problematic Adreno 5XX GPU configurations
- Removed duplicate isProblematicAdrenoGpu() method from VkThread.kt and updated
  to use the centralized method from DeviceUtils
- Removed duplicate isProblematicAdrenoGpuForVulkan() method from Godot.kt and
  updated all call sites to use DeviceUtils.isProblematicAdrenoGpu()
- Removed duplicate isProblematicAdrenoGpu() method from GodotGLRenderView.java
  and updated to use DeviceUtils.isProblematicAdrenoGpu()
- Cleaned up unused imports (BufferedReader, FileReader) from GodotGLRenderView.java

This refactoring improves code maintainability by eliminating code duplication
and centralizing GPU detection logic in a single location.
